+++ CMakeLists.txt
@@ -35,6 +35,7 @@ if(NANOOSC_BUILD_RUNTIME_STATS)
 	endif()
 	add_library(nanoosc_runtime_stats src/runtime_stats.cpp)
 	target_link_libraries(nanoosc_runtime_stats PUBLIC nanoosc::runtime)
+	target_compile_definitions(nanoosc_runtime_stats PUBLIC NANOOSC_ENABLE_RUNTIME_STATS=1)
 	target_compile_features(nanoosc_runtime_stats PUBLIC cxx_std_17)
 	add_library(nanoosc::runtime_stats ALIAS nanoosc_runtime_stats)
 endif()
@@ -48,8 +49,15 @@ if(NANOOSC_BUILD_UDP)
 	if(NOT NANOOSC_BUILD_RUNTIME)
 		message(FATAL_ERROR "NANOOSC_BUILD_UDP requires NANOOSC_BUILD_RUNTIME=ON")
 	endif()
+	if(WIN32)
+		message(FATAL_ERROR "NANOOSC_BUILD_UDP currently requires POSIX sockets; Windows UDP support is planned for v0.4.0")
+	endif()
 	add_library(nanoosc_udp src/udp_transport_posix.cpp)
 	target_link_libraries(nanoosc_udp PUBLIC nanoosc::runtime)
+	if(TARGET nanoosc::runtime_stats)
+		target_sources(nanoosc_udp PRIVATE src/udp_diagnostics_posix.cpp)
+		target_link_libraries(nanoosc_udp PUBLIC nanoosc::runtime_stats)
+	endif()
 	target_compile_features(nanoosc_udp PUBLIC cxx_std_17)
 	add_library(nanoosc::udp ALIAS nanoosc_udp)
 endif()
@@ -60,15 +68,12 @@ target_link_libraries(nanoosc INTERFACE nanoosc::core)
 if(TARGET nanoosc::runtime)
 	target_link_libraries(nanoosc INTERFACE nanoosc::runtime)
 endif()
-if(TARGET nanoosc::udp)
-	target_link_libraries(nanoosc INTERFACE nanoosc::udp)
-endif()
 # Namespaced alias
 add_library(nanoosc::nanoosc ALIAS nanoosc)
 
 option(BUILD_EXAMPLES "Build examples during the build" ON)
 if(BUILD_EXAMPLES)
-	if(TARGET nanoosc::udp OR TARGET nanoosc::runtime_stats)
+	if(TARGET nanoosc::udp)
 		add_subdirectory(examples)
 	else()
 		message(STATUS "Skipping examples because no example transport targets are enabled")
