+++ examples/CMakeLists.txt
@@ -29,12 +29,12 @@ function(add_nanoosc_example name)
     set_target_properties(${name} PROPERTIES FOLDER "examples")
 endfunction()
 
-function(add_nanoosc_runtime_stats_example name)
-    if(NOT TARGET nanoosc::runtime_stats)
-        message(FATAL_ERROR "${name} requires runtime stats. Configure with -DNANOOSC_BUILD_RUNTIME_STATS=ON")
+function(add_nanoosc_udp_stats_example name)
+    if(NOT TARGET nanoosc::udp OR NOT TARGET nanoosc::runtime_stats)
+        message(FATAL_ERROR "${name} requires UDP and runtime stats. Configure with -DNANOOSC_BUILD_UDP=ON -DNANOOSC_BUILD_RUNTIME_STATS=ON")
     endif()
     add_executable(${name} ${ARGN})
-    target_link_libraries(${name} PRIVATE nanoosc::runtime_stats)
+    target_link_libraries(${name} PRIVATE nanoosc::udp)
     set_target_properties(${name} PROPERTIES FOLDER "examples")
 endfunction()
 
@@ -49,8 +49,8 @@ else()
     message(STATUS "Skipping UDP examples because nanoosc::udp is not enabled")
 endif()
 
-if(TARGET nanoosc::runtime_stats)
+if(TARGET nanoosc::udp AND TARGET nanoosc::runtime_stats)
     add_subdirectory(osc-stats-server)
 else()
-    message(STATUS "Skipping OSCStatsServer example because NANOOSC_BUILD_RUNTIME_STATS=OFF")
+    message(STATUS "Skipping OSCStatsServer example because UDP or runtime stats are not enabled")
 endif()
