+++ examples/osc-stats-server/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_nanoosc_runtime_stats_example(osc-stats-server src/osc-stats-server.cpp)
+add_nanoosc_udp_stats_example(osc-stats-server src/osc-stats-server.cpp)
 
 add_dependencies(examples osc-stats-server)
